simple as a node app
