--- .eslintrc
@@ -1,29 +1,0 @@
-{
-  "extends": [
-    "@opengovsg/eslint-config-opengovsg"
-  ],
-  "parser": "babel-eslint",
-  "parserOptions": {
-    "ecmaVersion": 6,
-    "sourceType": "module"
-  },
-  "env": {
-    "node": true,
-    "es6": true,
-    "jasmine": true
-  },
-  "plugins": [
-    "jasmine",
-    "flowtype",
-    "flowtype-comment"
-  ],
-  "rules": {
-    "no-console": 1,
-    "no-underscore-dangle": 1
-  },
-  "settings": {
-    "flowtype": {
-      "onlyFilesWithFlowAnnotation": true
-    }
-  }
-}
