+++ .gitignore
@@ -1,9 +1,3 @@
 .DS_Store
-.next/
 node_modules/
-build/
-npm-debug.log
-.eslintcache
-dist/
-output/
-.nyc_output/
+package-lock.json
