--- .nycrc
@@ -1,9 +1,0 @@
-{
-  "all": true,
-  "include": [
-    "lib/**"
-  ],
-  "exclude": [
-    "test/**/*.spec.js"
-  ]
-}
