+++ swagger.json
@@ -1,0 +1,44 @@
+{
+    "openapi": "3.0.0",
+    "info": {
+      "version": "1.0.0",
+      "title": "PDF to Markdown API",
+      "description": "API to convert PDF to Markdown"
+    },
+    "servers": [
+      {
+        "url": "http://localhost:3001"
+      }
+    ],
+    "paths": {
+      "/api/convertPdfToMd": {
+        "post": {
+          "summary": "Convert PDF to Markdown",
+          "responses": {
+            "200": {
+              "description": "Successful conversion",
+              "content": {
+                "application/json": {
+                  "schema": {
+                    "type": "object",
+                    "properties": {
+                      "markdown": {
+                        "type": "string"
+                      }
+                    }
+                  }
+                }
+              }
+            },
+            "400": {
+              "description": "Bad Request"
+            },
+            "500": {
+              "description": "Internal Server Error"
+            }
+          }
+        }
+      }
+    }
+  }
+
